moving method funcs to utils
